+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 *   Added waveform hotkeys for zoom all and clear all markers #105
 *   Added path option as fixed, most recently, used or most recently saved #92
 *   Clear energy and charge on JLS file open and "disable" device #99
+*   Fixed multimeter value to work on units and names #89
 
 
 ## 0.9.2